Updating...
I'm pulling in my latest changes from my 9-5 job.  Don't worry, I have permission.  We're working on decoupling some of the basics so that we can make portions of the project open source, but for now, my boss just told me to consider the more generic and boring pieces as parts that I can contribute to other projects.